ai/requirements.txt
fastapi==0.115.0
Pillow==10.4.0
protobuf==5.28.2
pydantic==2.9.2
pymongo==4.9.1
Requests==2.32.3
setuptools==75.1.0
tavily_python==0.5.0
